vmware-horizon-client-np: add checkver, autoupdate
Add support for checkver and autoupdate to simplify maintenance of
vmware-horizon-client-np manifest. Have not found a central registry for
VMware Horizon Client versions, so the `downloadGroup` parameter in
`checkver.url` requires manual update.

Due to the manual update requirement for download group with each minor
release, unsure if this makes the `checkver` and `autoupdate` worth it.
If nothing else this helps prevent typos in version or hash.