barrier: setup barrier virtual kvm
Barrier is a fork of Synergy that focuses on stability and being simple
and free.

This installer is not completely hands-off. It requires permission
elevation because it needs to take control of your mouse (client) or add
a virtual monitor (server). The Barrier installer also asks to install
Bonjour (for discovery). It installs into Program Files. Seems like
nonportable is the right bucket.

This json is based on:
https://gist.github.com/pluser/239add9975080a01e2bb93b28e174acb

I changed the description and formatting.